rename the discovery package to mdns
