+++ examples/ipfs-camp-2019/06-Pubsub/main.go
@@ -17,7 +17,7 @@ import (
 	mplex "github.com/libp2p/go-libp2p-mplex"
 	tls "github.com/libp2p/go-libp2p-tls"
 	yamux "github.com/libp2p/go-libp2p-yamux"
-	"github.com/libp2p/go-libp2p/p2p/discovery"
+	"github.com/libp2p/go-libp2p/p2p/mdns"
 	"github.com/libp2p/go-tcp-transport"
 	ws "github.com/libp2p/go-ws-transport"
 	"github.com/multiformats/go-multiaddr"
@@ -101,7 +101,7 @@ func main() {
 		fmt.Println("Connected to", targetInfo.ID)
 	}
 
-	mdns := discovery.NewMdnsService(host, "")
+	mdns := mdns.NewMdnsService(host, "")
 	notifee := &discoveryNotifee{h: host, ctx: ctx}
 	mdns.RegisterNotifee(notifee)
