+++ examples/ipfs-camp-2019/08-End/main.go
@@ -16,7 +16,7 @@ import (
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	tls "github.com/libp2p/go-libp2p-tls"
 	yamux "github.com/libp2p/go-libp2p-yamux"
-	"github.com/libp2p/go-libp2p/p2p/discovery"
+	"github.com/libp2p/go-libp2p/p2p/mdns"
 	"github.com/libp2p/go-tcp-transport"
 	ws "github.com/libp2p/go-ws-transport"
 	"github.com/multiformats/go-multiaddr"
@@ -108,7 +108,7 @@ func main() {
 
 	fmt.Println("Connected to", targetInfo.ID)
 
-	mdns := discovery.NewMdnsService(host, "")
+	mdns := mdns.NewMdnsService(host, "")
 	mdns.RegisterNotifee(&mdnsNotifee{h: host, ctx: ctx})
 
 	err = dht.Bootstrap(ctx)
