+++ p2p/mdns/mdns.go
@@ -1,4 +1,4 @@
-package discovery
+package mdns
 
 import (
 	"context"
