+++ p2p/mdns/mdns_test.go
@@ -1,4 +1,4 @@
-package discovery
+package mdns
 
 import (
 	"context"
