+++ p2p/mdns_legacy/mdns.go
@@ -1,4 +1,4 @@
-package discovery_legacy
+package mdns_legacy
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/libp2p/go-libp2p/p2p/discovery"
+	mdnsnew "github.com/libp2p/go-libp2p/p2p/mdns"
 
 	logging "github.com/ipfs/go-log/v2"
 	ma "github.com/multiformats/go-multiaddr"
@@ -26,8 +26,8 @@ var log = logging.Logger("mdns_legacy")
 
 const ServiceTag = "_ipfs-discovery._udp"
 
-type Service = discovery.Service
-type Notifee = discovery.Notifee
+type Service = mdnsnew.Service
+type Notifee = mdnsnew.Notifee
 
 type mdnsService struct {
 	server  *mdns.Server
