+++ p2p/mdns_legacy/mdns_test.go
@@ -1,4 +1,4 @@
-package discovery_legacy
+package mdns_legacy
 
 import (
 	"context"
